Split Reply Comic Component
